requirements.txt
pygame~=2.5.2
pillow~=10.3.0